rename: 컨벤션에 맞도록 NotificationItem 컴포넌트 경로 수정
